Use u64 in virtual memory API

+++ include/dn/mem/virtual.h
@@ -1,6 +1,6 @@
 #pragma once
 
-u32 DnMemVirtualPageSize();
+u64 DnMemVirtualPageSize();
 void* DnMemVirtualReserve(u64 size);
 bool DnMemVirtualCommit(void* ptr, u64 size);
 void DnMemVirtualDecommit(void* ptr, u64 size);
